requirements.txt
sympy
google-generativeai
flask
flask-cors
